apps/api/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
pydantic==2.8.2
pydantic-settings==2.4.0
httpx==0.27.0
python-dotenv==1.0.1

# LangChain and LangGraph for Agents
langgraph==0.2.28
langchain==0.3.1
langchain-openai==0.2.1
langchain-community==0.3.1
langchain-core==0.3.6

# Database
sqlalchemy==2.0.35
asyncpg==0.29.0
alembic==1.13.3

# Image Processing
pillow==10.4.0
python-multipart==0.0.17

# AWS S3 (optional)
boto3==1.35.36

# Additional utilities
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
aiofiles==24.1.0

# Testing
pytest==8.3.3
pytest-asyncio==0.24.0
httpx==0.27.0